weather-dashboard/backend-python/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
apscheduler
groq
python-dotenv
fastapi-cors
